requirements.txt
google-genai
python-dotenv
opencv-python-headless
